model-downloader/requirements.txt
huggingface_hub
google-cloud-storage